// mobile-client/src/model/enum.ts
// 逃跑
export enum FlEE {
  YES = '1',
  NO = '0',
}

// 房间
export enum ROOM {
  // 普通场
  STANDARD = '0',
  // 至尊场
  ADVANCE = '1',
}

// 倍数
export enum MULTIPLE {
  ONE = '1',
  TWO = '2',
  THREE = '3',
}

// 玩家
export enum PLAYER {
  SELF = '0',
  REIVAL = '1',
}

// 结果
export enum RESULT {
  ALL = 'all',
  WIN = '0',
  LOSE = '1',
}

// 角色
export enum ROLE {
  ALL = 'all',
  LAND_LORD = '0',
  FARMER1 = '1',
  FARMER2 = '2',
}